dom/indexedDB/test/test_put_get_values_autoIncrement.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
    function testSteps()
    {
      const name = window.location.pathname;
      const description = "My Test Database";
      const objectStoreName = "Objects";

      let testString = { value: "testString" };
      let testInt = { value: 1002 };

      let request = moz_indexedDB.open(name, description);
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield;

      let db = event.result;

      request = db.setVersion("1");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      event = yield;

      let objectStore = db.createObjectStore(objectStoreName, "", true);

      request = objectStore.add(testString.value);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        testString.key = event.result;
        request = objectStore.get(testString.key);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          is(event.result, testString.value, "Got the right value");
        };
      };

      request = objectStore.add(testInt.value);
      request.onerror = errorHandler;
      request.onsuccess = function(event) {
        testInt.key = event.result;
        request = objectStore.get(testInt.key);
        request.onerror = errorHandler;
        request.onsuccess = function(event) {
          is(event.result, testInt.value, "Got the right value");
          finishTest();
        };
      }

      yield;
    }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
